check if band + .fits exists

+++ TDEpy/reduction.py
@@ -36,7 +36,9 @@ def do_sw_photo(sw_dir, aper_cor):
                 pass
             exists = os.path.isfile('sw' + str(dirs[i]) + 'u' + band + '_sk.img.gz')
             if exists:
-                os.system('rm ' + band + '.fits')
+                exists_old = os.path.isfile(band + '.fits')
+                if exists_old:
+                    os.system('rm ' + band + '.fits')
                 if aper_cor:
                     os.system('uvotsource image=sw' + str(dirs[i]) + 'u' + band + '_sk.img.gz srcreg=source.reg '
                                                                                   'bkgreg=bkg.reg sigma=3 '
